fix: Address code review feedback
- Fix critical navigation state persistence logic: Now stores 'expanded' state instead of 'collapsed'
- Refactor theme utility functions to eliminate code duplication: Export shared functions from themes.ts
