+++ PR-46-TEST-RESULTS.md
@@ -1,0 +1,103 @@
+# PR #46 Test Results - Ready for Merge ✅
+
+## Testing Summary
+
+All tests have been completed successfully after merging the latest main branch (which includes PR #44 and PR #45).
+
+### Merge Conflicts Resolution ✅
+
+- Successfully merged main branch into feature/simplify-theme-api
+- Resolved conflicts in:
+  - `src/theme-switcher.ts` - Preserved accessibility fix for focus management
+  - `src/viewer.ts` - Preserved search content indexing fix and navigation state persistence
+
+### Test Results
+
+#### 1. Test Suite ✅
+
+```bash
+npm test
+```
+
+- **Result**: All 367 tests pass
+- **Coverage**: Meets all thresholds (80%+)
+
+#### 2. Build Process ✅
+
+```bash
+npm run build
+```
+
+- **Result**: Build completed successfully
+- **Output**: ES and UMD formats generated
+- **Note**: Expected warning about "build.lib.formats" is non-critical
+
+#### 3. TypeScript Compilation ✅
+
+```bash
+npm run typecheck
+```
+
+- **Result**: No TypeScript errors
+
+#### 4. Linting ✅
+
+```bash
+npm run lint
+```
+
+- **Result**: All linting checks pass (fixed unused catch parameters)
+
+#### 5. Development Server ✅
+
+```bash
+npm run dev
+```
+
+- **Result**: Server starts successfully on port 5000
+
+#### 6. Demo Application ✅
+
+```bash
+npm run demo:build
+```
+
+- **Result**: Demo builds successfully
+
+### UI Improvements Verification ✅
+
+All UI improvements from this PR are working correctly:
+
+1. **Navigation Categories Arrow Direction** ✅
+   - Arrows point correctly (▶ collapsed, ▼ expanded)
+
+2. **Navigation State Persistence** ✅
+   - Category expand/collapse states saved to localStorage
+   - States restored when switching documents
+
+3. **Search Input Theme Transitions** ✅
+   - Smooth color transitions when theme changes
+
+4. **Search Content Indexing Fix** ✅
+   - Documents with inline content are properly indexed
+
+5. **Navigation-Content Spacing** ✅
+   - Reduced padding from 32px to 16px
+
+6. **Desktop Sidebar Collapse** ✅
+   - Toggle button visible on desktop
+   - Smooth width animation
+   - Ctrl+B keyboard shortcut works
+   - State persisted in localStorage
+   - Proper ARIA attributes
+
+### Environment
+
+- Node.js: v22.17.0
+- npm: 10.8.2
+- Platform: darwin
+- Vite: 7.0.6 (upgraded from 4.5.5)
+
+### Conclusion
+
+PR #46 is ready for merge. All UI improvements are functioning correctly, and the codebase is fully compatible with the latest dependencies including Vite 7.
