+++ src/theme-builder.ts
@@ -6,16 +6,7 @@ import {
   sanitizeFontFamily,
   sanitizeCssValue,
 } from './utils';
-import { createCustomTheme } from './themes';
-
-// Utility functions for theme management
-function getThemeBaseName(themeName: string): string {
-  return themeName.replace(/-(light|dark)$/, '');
-}
-
-function getThemeMode(themeName: string): 'light' | 'dark' {
-  return themeName.endsWith('-dark') ? 'dark' : 'light';
-}
+import { createCustomTheme, getThemeBaseName, getThemeMode } from './themes';
 
 type ThemeColors = Theme['colors'];
 type ThemeFonts = Theme['fonts'];
